100_mentor_PCSI/requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai==1.43.0
anthropic==0.40.0
google-generativeai==0.7.2
Pillow==10.4.0
chromadb>=0.5.5
pypdf>=4.2.0
pymupdf==1.26.3